[DRAFT] fix typo

+++ Libraries/Base/Source/Stream.cc
@@ -91,7 +91,7 @@ Stream& Stream::operator>>(String& message) {
 
 Stream& Stream::operator<<(Byte&& value) {
   if (_Writer) {
-    auto size_of_value = sizeof(value);
+    auto size_of_value = UInt(sizeof(value));
     auto error = _Writer((Bytes)&value, &size_of_value);
 
     if (error != ENoError) throw Exception(error);
@@ -101,7 +101,7 @@ Stream& Stream::operator<<(Byte&& value) {
 
 Stream& Stream::operator<<(Int&& value) {
   if (_Writer) {
-    auto size_of_value = sizeof(value);
+    auto size_of_value = UInt(sizeof(value));
     auto error = _Writer((Bytes)&value, &size_of_value);
 
     if (error != ENoError) throw Exception(error);
@@ -111,7 +111,7 @@ Stream& Stream::operator<<(Int&& value) {
 
 Stream& Stream::operator<<(UInt&& value) {
   if (_Writer) {
-    auto size_of_value = sizeof(value);
+    auto size_of_value = UInt(sizeof(value));
     auto error = _Writer((Bytes)&value, &size_of_value);
 
     if (error != ENoError) throw Exception(error);
@@ -121,7 +121,7 @@ Stream& Stream::operator<<(UInt&& value) {
 
 Stream& Stream::operator<<(Float&& value) {
   if (_Writer) {
-    auto size_of_value = sizeof(value);
+    auto size_of_value = UInt(sizeof(value));
     auto error = _Writer((Bytes)&value, &size_of_value);
 
     if (error != ENoError) throw Exception(error);
@@ -131,7 +131,7 @@ Stream& Stream::operator<<(Float&& value) {
 
 Stream& Stream::operator<<(Double&& value) {
   if (_Writer) {
-    auto size_of_value = sizeof(value);
+    auto size_of_value = UInt(sizeof(value));
     auto error = _Writer((Bytes)&value, &size_of_value);
 
     if (error != ENoError) throw Exception(error);
